shouldn't redirect to login/register page if not logged in

+++ src/main/java/project/controller/SheetController.java
@@ -39,6 +39,9 @@ public class SheetController {
         model.addAttribute("charbean", new CharacterBean());
 
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         if(user.getUserID() != null)
             return "newcharacter";
@@ -53,6 +56,9 @@ public class SheetController {
     public String newCharacterPost(@ModelAttribute CharacterBean charbean, Model model, HttpSession session) {
         // Make sure user is who we think they are, and put character bean in our model.
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
 
 
@@ -92,6 +98,9 @@ public class SheetController {
         model.addAttribute("charbean", new CharacterBean());
 
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         if(user.getUserID() != null)
             return "newcharacter";
@@ -105,6 +114,9 @@ public class SheetController {
     public String newCharacterCampaignPost(@ModelAttribute CharacterBean charbean, Model model, HttpSession session, @PathVariable int campID) {
         // Make sure user is who we think they are, and put character bean in our model.
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
 
 
@@ -137,6 +149,9 @@ public class SheetController {
     @RequestMapping(value = "/character{charID}", method = RequestMethod.GET)
     public String openCharacter(@PathVariable int charID, Model model, HttpSession session) {
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         storage = new AccountStorage("data/userAccounts.sqlite");
         model.addAttribute("myChars", storage.listCharacters(user.getUserID()));
@@ -161,6 +176,9 @@ public class SheetController {
     @RequestMapping(value = "/updateCharacter", method = RequestMethod.POST)
     public String myCharacterPost(@ModelAttribute CharacterBean charbean, Model model, HttpSession session) {
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         CharacterBean oldBean = (CharacterBean)session.getAttribute("charbean");
 
         charbean.setSpellSlots_details(oldBean.getSpellSlots_details());
@@ -186,6 +204,9 @@ public class SheetController {
         System.out.println("leveling up character "+charID+" one level in class "+classID);
 
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         storage = new AccountStorage("data/userAccounts.sqlite");
         Lookup find = new Lookup();
@@ -222,6 +243,9 @@ public class SheetController {
         System.out.println("ZZZ...");
 
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         storage = new AccountStorage("data/userAccounts.sqlite");
         Lookup find = new Lookup();
@@ -254,6 +278,9 @@ public class SheetController {
     @RequestMapping(value = "deleteCharacter{charID}", method = RequestMethod.GET)
     public String deleteCharacter(@PathVariable int charID, Model model, HttpSession session) {
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
         storage = new AccountStorage("data/userAccounts.sqlite");
         storage.deleteCharacter(charID);
@@ -269,6 +296,9 @@ public class SheetController {
     @RequestMapping(value = "/updateSpellslot", method = RequestMethod.POST)
     public String updateSpellslots(@RequestParam Map<String, String> allRequestParams, @ModelAttribute CharacterBean charbean, Model model, HttpSession session) {
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         CharacterBean oldBean = (CharacterBean)session.getAttribute("charbean");
         charbean.setSpellSlots_details(oldBean.getSpellSlots_details());
 
@@ -317,6 +347,9 @@ public class SheetController {
     public String listCharacters(Model model, HttpSession session) {
 
         User user = (User)session.getAttribute("userId");
+        if (user == null) {
+            return "redirect:/";
+        }
         model.addAttribute("user", user);
 
         storage = new AccountStorage("data/userAccounts.sqlite");
